fix #141159 (#143909)
